added amplicon-specific code (see especially emirge_amplicon.py)
This is a major update to EMIRGE that adds the ability to run emirge on amplicons generated from near-full-length genes.  The script emirge_amplicon.py should be able to handle large numbers of reads (up to a few million) that are all generated from the gene of interest (e.g. the 16S rRNA gene).  There are several space for speed tradeoffs, much more efficient code, etc.  There is a new Cython file as well.  Eventually, the two code bases will be merged.

This version of the code (specifically, emirge_amplicon.py) was used for the manuscript describing the use of EMIRGE on amplicon data (submitted).

Additional changes:
  changed utils/fix_nonstandard_chars.py to remove any gap characters (- or .)
  added pykseq, a Cython wrapper of Heng Li's fast sequence parser: kseq.h
  updated README