Added ability to resume iterations for emirge.py
This functionality (--resume_from) is now documented in --help.  It requires that completed directories for a finished iteration and the previous iteration are present.  It should be exactly as if you had paused and resumed the previous run.  This also allows for continuing a run beyond the initial number of iterations specified.

You should probably keep all other command line options the same.  You can grep your original command out of the emirge output like this:
grep -A 1 Command emirge.log

Adding this functionality to emirge_amplicon.py might be a little more non-trivial.  For now, only option is to restart using as the database the last finished EMIRGE iteration.