Fix compatability with BioPython >= 1.26 (fixes #26)
Since v1.26, BioPython SeqRecord raises NotImplemented exceptions in all
comparator methods. The list of (prior,SeqRecord) pairs should have a
total order based on the prior alone, but Python's Timsort seems to
make calls to ==, which get passed to the second item in the tuples and
trigger the exception in SeqRecord.__eq__. Sorting purely on the priors
by passing a comparator fixes the issue.